move remaining pattern functions
